requirements.txt
python-telegram-bot==21.0.1
APScheduler==3.10.4
PyYAML==6.0.1
watchdog